# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Continuity API - re-entry briefing app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  coasting: "coasting",
  dark: "dark",
} as const;

export type ProjectConfidenceLevel =
  (typeof ProjectConfidenceLevel)[keyof typeof ProjectConfidenceLevel];

export const ProjectConfidenceLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type ProjectProjectType =
  (typeof ProjectProjectType)[keyof typeof ProjectProjectType];

export const ProjectProjectType = {
  developer: "developer",
  smb: "smb",
  creative: "creative",
  other: "other",
} as const;

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  confidenceLevel: ProjectConfidenceLevel;
  projectType: ProjectProjectType;
  /** @nullable */
  lastActivityAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ProjectDetailStatus =
  (typeof ProjectDetailStatus)[keyof typeof ProjectDetailStatus];

export const ProjectDetailStatus = {
  active: "active",
  coasting: "coasting",
  dark: "dark",
} as const;

export type ProjectDetailConfidenceLevel =
  (typeof ProjectDetailConfidenceLevel)[keyof typeof ProjectDetailConfidenceLevel];

export const ProjectDetailConfidenceLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type ProjectDetailProjectType =
  (typeof ProjectDetailProjectType)[keyof typeof ProjectDetailProjectType];

export const ProjectDetailProjectType = {
  developer: "developer",
  smb: "smb",
  creative: "creative",
  other: "other",
} as const;

export type BriefingConfidenceLevel =
  (typeof BriefingConfidenceLevel)[keyof typeof BriefingConfidenceLevel];

export const BriefingConfidenceLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Briefing {
  id: number;
  projectId: number;
  lastKnownState: string;
  confidenceLevel: BriefingConfidenceLevel;
  blockers: string[];
  nextActions: string[];
  rawOutput: string;
  createdAt: string;
}

export type UpdateSourceType =
  (typeof UpdateSourceType)[keyof typeof UpdateSourceType];

export const UpdateSourceType = {
  text: "text",
  voice: "voice",
  worksheet: "worksheet",
} as const;

export interface Update {
  id: number;
  projectId: number;
  content: string;
  sourceType: UpdateSourceType;
  tags: string[];
  createdAt: string;
}

export interface ProjectDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ProjectDetailStatus;
  confidenceLevel: ProjectDetailConfidenceLevel;
  projectType: ProjectDetailProjectType;
  /** @nullable */
  lastActivityAt?: string | null;
  createdAt: string;
  updatedAt: string;
  latestBriefing?: Briefing;
  recentUpdates: Update[];
}

export type CreateProjectBodyProjectType =
  (typeof CreateProjectBodyProjectType)[keyof typeof CreateProjectBodyProjectType];

export const CreateProjectBodyProjectType = {
  developer: "developer",
  smb: "smb",
  creative: "creative",
  other: "other",
} as const;

export interface CreateProjectBody {
  title: string;
  description?: string;
  projectType?: CreateProjectBodyProjectType;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  active: "active",
  coasting: "coasting",
  dark: "dark",
} as const;

export type UpdateProjectBodyProjectType =
  (typeof UpdateProjectBodyProjectType)[keyof typeof UpdateProjectBodyProjectType];

export const UpdateProjectBodyProjectType = {
  developer: "developer",
  smb: "smb",
  creative: "creative",
  other: "other",
} as const;

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  status?: UpdateProjectBodyStatus;
  projectType?: UpdateProjectBodyProjectType;
}

export type CreateUpdateBodySourceType =
  (typeof CreateUpdateBodySourceType)[keyof typeof CreateUpdateBodySourceType];

export const CreateUpdateBodySourceType = {
  text: "text",
  voice: "voice",
  worksheet: "worksheet",
} as const;

export interface CreateUpdateBody {
  content: string;
  sourceType?: CreateUpdateBodySourceType;
}

export interface SubmitWorksheetBody {
  momentumRating: number;
  currentStatus: string;
  /** @nullable */
  mainBlocker?: string | null;
  nextStep: string;
  sameAsLast?: boolean;
}

export interface DashboardStats {
  totalProjects: number;
  activeProjects: number;
  coastingProjects: number;
  darkProjects: number;
  totalUpdates: number;
  totalBriefings: number;
  recentActivity: Update[];
}