# lib/api-zod/src/generated/types/briefing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Continuity API - re-entry briefing app
 * OpenAPI spec version: 0.1.0
 */
import type { BriefingConfidenceLevel } from "./briefingConfidenceLevel";

export interface Briefing {
  id: number;
  projectId: number;
  lastKnownState: string;
  confidenceLevel: BriefingConfidenceLevel;
  blockers: string[];
  nextActions: string[];
  rawOutput: string;
  createdAt: string;
}