# lib/api-zod/src/generated/types/createUpdateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Continuity API - re-entry briefing app
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUpdateBodySourceType } from "./createUpdateBodySourceType";

export interface CreateUpdateBody {
  content: string;
  sourceType?: CreateUpdateBodySourceType;
}