# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Continuity API - re-entry briefing app
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectConfidenceLevel } from "./projectConfidenceLevel";
import type { ProjectProjectType } from "./projectProjectType";
import type { ProjectStatus } from "./projectStatus";

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  confidenceLevel: ProjectConfidenceLevel;
  projectType: ProjectProjectType;
  /** @nullable */
  lastActivityAt?: string | null;
  createdAt: string;
  updatedAt: string;
}