# lib/api-zod/src/generated/types/submitWorksheetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Continuity API - re-entry briefing app
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitWorksheetBody {
  momentumRating: number;
  currentStatus: string;
  /** @nullable */
  mainBlocker?: string | null;
  nextStep: string;
  sameAsLast?: boolean;
}