# lib/api-zod/src/generated/types/update.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Continuity API - re-entry briefing app
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSourceType } from "./updateSourceType";

export interface Update {
  id: number;
  projectId: number;
  content: string;
  sourceType: UpdateSourceType;
  tags: string[];
  createdAt: string;
}