# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Continuity API - re-entry briefing app
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyProjectType } from "./updateProjectBodyProjectType";
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  status?: UpdateProjectBodyStatus;
  projectType?: UpdateProjectBodyProjectType;
}